ci: clean up after Test Package stages if possible.
Basically, don't fail stage if jenkins was unable to checkout the git repo

+++ vars/runJenkinsPipeline.groovy
@@ -723,10 +723,12 @@ def call(){
                                                             }
                                                         }
                                                     } finally{
-                                                        if(isUnix()){
-                                                            sh "${tool(name: 'Default', type: 'git')} clean -dfx"
-                                                        } else {
-                                                            bat "${tool(name: 'Default', type: 'git')} clean -dfx"
+                                                        if(fileExists('.git')){
+                                                            if(isUnix()){
+                                                                sh "${tool(name: 'Default', type: 'git')} clean -dfx"
+                                                            } else {
+                                                                bat "${tool(name: 'Default', type: 'git')} clean -dfx"
+                                                            }
                                                         }
                                                     }
                                                 }
